requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
scikit-learn>=1.4.0
rich>=13.7.0
reportlab>=4.0.0
